refactor: 콘솔 관련 유틸및 유효성검사 함수를 LottoConsole로 이동
